requirements.txt
Django==5.2.11
python-dotenv==1.2.1
